refactor: ♻️ remove responsibility from main
