+++ lib/main.dart
@@ -1,29 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:parking_manager/shared/app_colors.dart';
-import 'package:parking_manager/shared/injector.dart';
-import 'package:parking_manager/shared/routes.dart';
+import 'package:parking_manager/my_app.dart';
+import 'package:parking_manager/injector.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
   await initDependencies();
   runApp(const MyApp());
-}
-
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp.router(
-      debugShowCheckedModeBanner: false,
-      title: 'Parking Manager',
-      theme: ThemeData(
-        colorScheme: ColorScheme.fromSeed(seedColor: AppColors.primary),
-        useMaterial3: true,
-      ),
-      routeInformationProvider: AppRouter.router.routeInformationProvider,
-      routeInformationParser: AppRouter.router.routeInformationParser,
-      routerDelegate: AppRouter.router.routerDelegate,
-    );
-  }
 }
