+++ lib/my_app.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+import 'package:parking_manager/shared/routes/routes.dart';
+import 'package:parking_manager/shared/themes/app_colors.dart';
+
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp.router(
+      debugShowCheckedModeBanner: false,
+      title: 'Parking Manager',
+      theme: ThemeData(
+        colorScheme: ColorScheme.fromSeed(seedColor: AppColors.primary),
+        useMaterial3: true,
+      ),
+      routeInformationProvider: AppRouter.router.routeInformationProvider,
+      routeInformationParser: AppRouter.router.routeInformationParser,
+      routerDelegate: AppRouter.router.routerDelegate,
+    );
+  }
+}
